Deque import rearragement
